CM-PIUG/requirements.txt
# CM-PIUG Dependencies
# ====================

# Core
numpy>=1.20.0
pyyaml>=5.4
dataclasses;python_version<"3.7"
typing-extensions>=3.7

# OCR
easyocr>=1.6.0
# pytesseract>=0.3.10  # Optional: apt install tesseract-ocr

# ASR
openai-whisper>=20230314
# SpeechRecognition>=3.8.0  # Optional fallback

# Machine Learning
transformers>=4.30.0
datasets>=2.14.0
accelerate>=0.26.0
torch>=2.0.0
sentence-transformers>=2.2.0  # Optional for semantic detection

# Image Processing
Pillow>=9.0.0

# Utilities
tqdm>=4.60.0
requests>=2.25.0

# API Server (optional)
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6

# Testing
pytest>=7.0.0
